Initial commit and open-source of Sinatra app that displays shuttle departure times and office locations.
It is generic in that if a CSV file is provided
that matches the provided CSV file columns, the 
HTML for the app can be regenerated.

It also serves as an example of using browser-based
geolocation and distance calculation to find the 
closest point of interest from ones provided, to the
user's browser-provided location. The distance calculation
is done in javascript.